fixup from pwndbg.aglib import regs doc
its not really a copy of the object, in this case `regs`
would be bound to the same object as `aglib.regs`
(None) and changes in one would reflect the other
if it wasn't None.

It is a kind of "copy of the name", which is what
you were going for, but maybe this edit makes it
a bit less ambiguous.